extracted lines processing to another file
